# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A production-ready geospatial routing microservice. Supports self-hosted OSRM (default) with an automatic in-process A* + Yen's k-shortest-paths fallback, live GPS tracking, WebSocket event delivery, and a PostGIS-backed OSM road graph.",
        "title": "geo-service",
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/Amirbehzad11/geo-service/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/route/waypoints": {
            "post": {
                "description": "Accepts a waypoint list (≥ 2) and a transport mode.\nKeeps the first waypoint as the start, then reorders the rest\nby nearest-neighbor (Haversine distance) before routing each\nconsecutive pair concurrently and stitching the polylines.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routing"
                ],
                "summary": "Calculate a multi-stop route",
                "parameters": [
                    {
                        "description": "Waypoints + mode",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_route.MultiRouteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/geo-service_internal_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_route.MultiRouteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    }
                }
            }
        },
...
        "/ws/trip/{id}": {
            "get": {
                "description": "Upgrades the connection to WebSocket and streams real-time `location.updated` and `deviation.detected` events for the specified trip via Redis Pub/Sub. Send any message to keep the connection alive.",
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket — live trip events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols — WebSocket upgrade successful"
                    },
                    "400": {
                        "description": "Missing or invalid trip ID",
                        "schema": {
                            "$ref": "#/definitions/geo-service_internal_response.Failure"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "geo-service_internal_response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "geo-service_internal_response.Failure": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/geo-service_internal_response.ErrorDetail"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "geo-service_internal_response.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_route.MultiRouteLeg": {
            "type": "object",
            "properties": {
                "distance_km": {
                    "type": "number"
                },
                "duration_min": {
                    "type": "number"
                },
                "from": {
                    "$ref": "#/definitions/internal_route.Waypoint"
                },
                "to": {
                    "$ref": "#/definitions/internal_route.Waypoint"
                }
            }
        },
        "internal_route.MultiRouteRequest": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "integer"
                },
                "waypoints": {
                    "description": "at least 2; first is the start",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_route.Waypoint"
                    }
                }
            }
        },
        "internal_route.MultiRouteResponse": {
            "type": "object",
            "properties": {
                "legs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_route.MultiRouteLeg"
                    }
                },
                "mode": {
                    "type": "string"
                },
                "polyline": {
                    "description": "combined path",
                    "type": "string"
                },
                "total_distance_km": {
                    "type": "number"
                },
                "total_duration_min": {
                    "type": "number"
                }
            }
        },
...
        "internal_route.Waypoint": {
            "type": "object",
            "properties": {
                "label": {
                    "description": "optional human-readable name",
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        }
    }
}